2-07-1: implement fps counting, fix a small bug.
FPS counting is implemented, but not displayed yet. It's an immediate fps, not a mean value. Also timespan is not checked for being to small and fps is not checked for being too high.

Fixed a bug. I wasn't destroying texture at a successful termination. Not a serious bug, but should be done.